chore: clean up build and auidt errors
