+++ package.json
@@ -93,7 +93,7 @@
     "eslint-config-next": "16.1.6",
     "husky": "^9.1.7",
     "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
+    "jest-environment-jsdom": "^30.2.0",
     "lint-staged": "^15.2.11",
     "next-sitemap": "^4.2.3",
     "prettier": "^3.4.1",
